fix(installer): cast DWMWA_USE_IMMERSIVE_DARK_MODE to u32 (i32 in windows-sys 0.59)

+++ desktop-installer/src-tauri/src/lib.rs
@@ -109,7 +109,7 @@ fn enable_dark_titlebar(win: &tauri::WebviewWindow) {
         unsafe {
             let _ = DwmSetWindowAttribute(
                 hwnd,
-                DWMWA_USE_IMMERSIVE_DARK_MODE,
+                DWMWA_USE_IMMERSIVE_DARK_MODE as u32,
                 &value as *const _ as *const _,
                 std::mem::size_of::<i32>() as u32,
             );
